Fix paging3 constant jumping issue
After enabling placeholders and jumping support, there was a strange issue after scrolling to a point in the list, where normally paging data would be invalidated and loading would start from 'scratch' from that point in this 'infinite' list. The issue was that after scrolling to that point, paging would repeatedly call `getRefreshKey` with improper anchor position values which also followed some king of pattern (e.g. 100, 250, 100, 50, 100, 250...). This would trigger the loading of the page(s) that were supposed to have the data for those positions, non-stop, resulting into an incorrect and 'never settling' list state.

After spending an hour debugging what was going wrong, turned out that latest Paging3 'stable' version had an issue calculating the anchor position, which was potentially (and actually) fixed in Paging 3.2.0+