Goodbye Paging3
Implemented simple pagination without paging3. A major issue with the Paging library was the inability to restore the scroll position of the list when navigating back to the list fragment or re-entering the app from background. Restoring scroll position would only work when placeholders were enabled, which is infeasible to do with dynamic headers and non same-height items.

With the new way of exposing data from the repository, the cache is also useless at the moment so it was removed.